function/bq_taxonomy/requirements.txt
anyio==4.11.0 ; python_version >= "3.12" and python_version < "4"
blinker==1.9.0 ; python_version >= "3.12" and python_version < "4"
cachetools==6.2.2 ; python_version >= "3.12" and python_version < "4.0"
certifi==2025.11.12 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.4 ; python_version >= "3.12" and python_version < "4.0"
click==8.3.0 ; python_version >= "3.12" and python_version < "4"
cloudevents==1.12.0 ; python_version >= "3.12" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4" and platform_system == "Windows"
deprecation==2.1.0 ; python_version >= "3.12" and python_version < "4"
flask==3.1.2 ; python_version >= "3.12" and python_version < "4"
functions-framework==3.10.0 ; python_version >= "3.12" and python_version < "4"
google-api-core==2.28.1 ; python_version >= "3.12" and python_version < "4.0"
google-api-core[grpc]==2.28.1 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.43.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-appengine-logging==1.7.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-audit-log==0.4.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-bigquery==3.38.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-core==2.5.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-dataform==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-logging==3.12.1 ; python_version >= "3.12" and python_version < "4.0"
google-crc32c==1.7.1 ; python_version >= "3.12" and python_version < "4.0"
google-resumable-media==2.7.2 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.72.0 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos[grpc]==1.72.0 ; python_version >= "3.12" and python_version < "4.0"
grpc-google-iam-v1==0.14.3 ; python_version >= "3.12" and python_version < "4.0"
grpcio-status==1.76.0 ; python_version >= "3.12" and python_version < "4.0"
grpcio==1.76.0 ; python_version >= "3.12" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.12" and python_version < "4"
h11==0.16.0 ; python_version >= "3.12" and python_version < "4"
idna==3.11 ; python_version >= "3.12" and python_version < "4.0"
importlib-metadata==8.7.0 ; python_version >= "3.12" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.12" and python_version < "4"
jinja2==3.1.6 ; python_version >= "3.12" and python_version < "4"
markupsafe==3.0.3 ; python_version >= "3.12" and python_version < "4"
opentelemetry-api==1.38.0 ; python_version >= "3.12" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.12" and python_version < "4.0"
proto-plus==1.26.1 ; python_version >= "3.12" and python_version < "4.0"
protobuf==6.33.1 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.2 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.5 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9.1 ; python_version >= "3.12" and python_version < "4"
six==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4"
starlette==0.50.0 ; python_version >= "3.12" and python_version < "4"
typing-extensions==4.15.0 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.5.0 ; python_version >= "3.12" and python_version < "4.0"
uvicorn-worker==0.4.0 ; python_version >= "3.12" and python_version < "4"
uvicorn==0.38.0 ; python_version >= "3.12" and python_version < "4"
watchdog==6.0.0 ; python_version >= "3.12" and python_version < "4"
werkzeug==3.1.3 ; python_version >= "3.12" and python_version < "4"
zipp==3.23.0 ; python_version >= "3.12" and python_version < "4.0"
